fixes circular dependency
