+++ packages/insomnia-app/app/common/misc.ts
@@ -466,3 +466,14 @@ export function isNotNullOrUndefined<ValueType>(
 }
 
 export const kebabCase = (value: string) => value.replace(/ /g, '-');
+
+// Because node-libcurl changed some names that we used in the timeline
+export const LIBCURL_DEBUG_MIGRATION_MAP = {
+  HeaderIn: 'HEADER_IN',
+  DataIn: 'DATA_IN',
+  SslDataIn: 'SSL_DATA_IN',
+  HeaderOut: 'HEADER_OUT',
+  DataOut: 'DATA_OUT',
+  SslDataOut: 'SSL_DATA_OUT',
+  Text: 'TEXT',
+} as const;
